Level 1 done, level 2 glitchy
I was able to add a new movement because the original one was very messed up and had no area to be able to add health damage taken or dealing with other mobs, I have to completely overall the player movement because of that. I was able to add other mobs and a boss that is just a different animation, coded a health, attack for mob and character, damage receive and dealt, a patrol which if they touch you they do damage and fixed the level a little.